+++ code/system.ui/react.media.video/src/ui/ui.VideoPlayer/types.mts
@@ -1,5 +1,5 @@
 import { type t } from './common';
 
-export type RootProps = {
+export type VideoPlayerProps = {
   style?: t.CssValue;
 };
